+++ src/containers/Accounts/index.tsx
@@ -1,5 +1,7 @@
-import { useEffect, useState } from 'react'
+import { useContext, useEffect, useState } from 'react'
 import { Helmet } from 'react-helmet-async'
+import { useQuery } from 'react-query'
+import { isValidClassicAddress, isValidXAddress } from 'ripple-address-codec'
 import { AccountHeader } from './AccountHeader'
 import { AccountTransactionTable } from './AccountTransactionTable'
 import './styles.scss'
@@ -8,6 +10,10 @@ import { Tabs } from '../shared/components/Tabs'
 import { AccountAssetTab } from './AccountAssetTab/AccountAssetTab'
 import { buildPath, useRouteParams } from '../shared/routing'
 import { ACCOUNT_ROUTE } from '../App/routes'
+import { BAD_REQUEST } from '../shared/utils'
+import { getAccountState } from '../../rippled'
+import SocketContext from '../shared/SocketContext'
+import { Loader } from '../shared/components/Loader'
 
 export const Accounts = () => {
   const { trackScreenLoaded } = useAnalytics()
@@ -15,6 +21,29 @@ export const Accounts = () => {
     useRouteParams(ACCOUNT_ROUTE)
   const [currencySelected, setCurrencySelected] = useState('XRP')
   const mainPath = buildPath(ACCOUNT_ROUTE, { id: accountId })
+  const rippledSocket = useContext(SocketContext)
+  const { trackException } = useAnalytics()
+
+  const { data: account, isLoading } = useQuery(
+    ['accountState', accountId],
+    () => {
+      if (!isValidClassicAddress(accountId) && !isValidXAddress(accountId)) {
+        return Promise.reject(BAD_REQUEST)
+      }
+
+      return getAccountState(accountId, rippledSocket).catch(
+        (transactionRequestError) => {
+          const status = transactionRequestError.code
+          trackException(
+            `ledger ${accountId} --- ${JSON.stringify(
+              transactionRequestError,
+            )}`,
+          )
+          return Promise.reject(status)
+        },
+      )
+    },
+  )
 
   useEffect(() => {
     trackScreenLoaded()
@@ -29,9 +58,10 @@ export const Accounts = () => {
   return (
     <div className="accounts-page section">
       <Helmet title={`${accountId.substring(0, 12)}...`} />
-      {accountId && (
+      {account && (
         <>
           <AccountHeader
+            account={account}
             accountId={accountId}
             onSetCurrencySelected={(currency) => setCurrencySelected(currency)}
             currencySelected={currencySelected}
@@ -44,9 +74,10 @@ export const Accounts = () => {
               hasTokensColumn={false}
             />
           )}
-          {tab === 'assets' && <AccountAssetTab account={accountId} />}
+          {tab === 'assets' && <AccountAssetTab account={account} />}
         </>
       )}
+      {isLoading && <Loader />}
     </div>
   )
 }
